feature(Pet): ペットクラスを作成してペットの共通処理をまとめた
ペットの共通のfunctionをPetクラスにまとめた

 #59
